+++ templates/sections/general.php
@@ -228,4 +228,5 @@
             <input type="hidden" name="settingkeykey" value="401">
         </div>
     </div>
+    
 </div>